Added implementation of POSIX signal set operations.
The implementation assumes that all supported signals have a value
less than 32. If the system does not have the 'sigset_t' type, it
is defined to an unsigned integer. We assume this is atleast 32-bits
wide as that is required by POSIX. For systems supporting larger
signal numbers, it is assumed that these functions and types are
provided. For example, GNU libc and modern BSD's use a struct
containing an array of unsigned integers. This implementation is
mainly for MinGW where NSIG is defined to 23.