Optimize 'strsep' and add tests.
Optimize 'strsep' to use more efficient functions that are typically
provided. This function is from 4.4BSD and is not standardized so it
may be missing on many systems. Use 'strchr' and 'strpbrk' to
implement this which are required by C99. Assume that the system
provides hand-written assembly implementations that are faster to call
than do ourselves.