Improve tests for 'unsetenv' and add a basic implementation.
Implementation has 2 code paths.
The first code path is for systems which declare 'unsetenv' but with
an old prototype that returns no value. In this case just remove the
variable until 'getenv' returns NULL.
In the second code path scan through 'environ' and move subsequent
elements backwards to remove the given variable.